refactor(napi/parser): move generated JS into `src-js` (#16957)
moves the generated code into src-js to keep the same pattern as oxlint
