+++ napi/parser/package.json
@@ -22,19 +22,19 @@
   "main": "src-js/index.js",
   "browser": "src-js/wasm.js",
   "files": [
-    "generated/constants.js",
-    "generated/deserialize/js.js",
-    "generated/deserialize/js_range.js",
-    "generated/deserialize/ts.js",
-    "generated/deserialize/ts_range.js",
-    "generated/lazy/constructors.js",
-    "generated/lazy/type_ids.js",
-    "generated/lazy/walk.js",
-    "generated/visit/keys.js",
-    "generated/visit/type_ids.js",
-    "generated/visit/visitor.d.ts",
-    "generated/visit/walk.js",
     "src-js/bindings.js",
+    "src-js/generated/constants.js",
+    "src-js/generated/deserialize/js.js",
+    "src-js/generated/deserialize/js_range.js",
+    "src-js/generated/deserialize/ts.js",
+    "src-js/generated/deserialize/ts_range.js",
+    "src-js/generated/lazy/constructors.js",
+    "src-js/generated/lazy/type_ids.js",
+    "src-js/generated/lazy/walk.js",
+    "src-js/generated/visit/keys.js",
+    "src-js/generated/visit/type_ids.js",
+    "src-js/generated/visit/visitor.d.ts",
+    "src-js/generated/visit/walk.js",
     "src-js/index.d.ts",
     "src-js/index.js",
     "src-js/raw-transfer/common.js",
